==> ETL/2.ETL_users_items.ipynb <==
ETL (Extracción, Transformación y Carga) es un proceso en el que se extraen datos de una fuente, se transforman y se cargan en otra fuente. En este caso, se extraen datos de un archivo json, se transforman y se cargan en un archivo csv.

# Cargamos las librerias necesarias para el desarrollo del ETL.
import sys
sys.path.append("../Lib")
import utils as ut
import pandas as pd
import gzip
import ast
import warnings
warnings.filterwarnings("ignore")

Creamos una función que se encargue de leer el archivo json y devolver un dataframe de pandas.

# Funcion para convertir un archivo .gz a un dataframe.
def gzip_to_df(file_path):                                          
    with gzip.open(file_path, 'rt', encoding='utf-8-sig') as file:  # Abrimos el archivo .gz
        return [ast.literal_eval(line) for line in file]            # Retornamos el archivo .gz como un dataframe.

# Cargamos los datos y los convertimos en un dataframe.
data = gzip_to_df('../Datasets/Raw/users_items.json.gz')     # Cargamos los datos.
df_users_items = pd.DataFrame(data)                                    # Convertimos los datos en un dataframe.
df_users_items.head()                                                  # Mostramos los primeros 5 registros del dataframe.

ut.data_summ(df_users_items)

df_users_items = pd.json_normalize(data, record_path=['items'], meta=['user_id','items_count','steam_id','user_url'] )
df_users_items

ut.data_summ(df_users_items)

# Obteniendo datos duplicados
duplicated = df_users_items.loc[df_users_items.duplicated()]
duplicated

# Eliminando datos duplicados
df_users_items.drop_duplicates(inplace=True)
df_users_items


# Eliminamos columnas que no se van a utilizar
df_users_items.drop(['playtime_2weeks', 'steam_id','user_url'], axis=1, inplace=True)
df_users_items

# Revisando si hay valores menores o iguales a 0 en la columna "playtime_forever"
df_users_items[df_users_items['playtime_forever'] <= 0]

Dado que en etapas posteriores utilizaremos la columna 'playtime_forever' para analizar el tiempo de juego, eliminaremos los registros con valores de 0, ya que no son relevantes para el análisis

df_users_items = df_users_items[df_users_items['playtime_forever'] > 0]
df_users_items

df_users_items.info()

df_users_items

# Guardar el dataframe en un archivo csv en la carpeta Clean
df_users_items.to_csv('../Datasets/Clean/users_items.csv.gz', compression='gzip', index=False, encoding='utf-8')

# Guardar el dataframe en un archivo parquet en la carpeta Datasets
df_users_items.to_parquet('../Datasets/users_items.parquet.gz', compression='gzip', index=False)
