==> ETL/3.ETL_user_reviews.ipynb <==
ETL (Extracción, Transformación y Carga) es un proceso en el que se extraen datos de una fuente, se transforman y se cargan en otra fuente. En este caso, se extraen datos de un archivo json, se transforman y se cargan en un archivo csv.

# Cargamos las librerias necesarias para el desarrollo del ETL.
import sys
sys.path.append("../Lib")
import utils as ut
import pandas as pd
import gzip
import ast
from dateutil import parser

Creamos una función que se encargue de leer el archivo json y devolver un dataframe de pandas.

# Funcion para convertir un archivo .gz a un dataframe.
def gzip_to_df(file_path):                                          
    with gzip.open(file_path, 'rt', encoding='utf-8-sig') as file:  # Abrimos el archivo .gz
        return [ast.literal_eval(line) for line in file]            # Retornamos el archivo .gz como un dataframe.

Iniciamos con el procesamientos de los datos de user_reviews.json. Para ello, hacemos uso de la función creada anteriormente y la guardamos en una variable llamada df_reviews.

# Cargamos los datos de los reviews y los convertimos en un dataframe.
data = gzip_to_df('../Datasets/Raw/user_Reviews.json.gz')     # Cargamos los datos de los reviews.
df_reviews = pd.DataFrame(data)                                    # Convertimos los datos de los reviews en un dataframe.
df_reviews.head()                                                  # Mostramos los primeros 5 registros del dataframe.

df_reviews.info() # Vemos la informacion del dataframe.

Vemos que la columna 'reviews' contiene un diccionario con los datos de cada review. Para poder acceder a los datos de cada review, creamos una función que se encargue de extraer los datos de cada review y devolver un dataframe de pandas.

# Desanidamos los datos de la columna reviews.
def extraer_values(df, columna):
    # Lista para almacenar los datos desanidados
    extracted_rows = []

    for index, row in df.iterrows(): # Itera sobre cada fila del DataFrame
        user_id = row["user_id"] 
        columnai = row[columna]
        
        if isinstance(columnai, list):
            for item in columnai:
                extracted_row = {"user_id": user_id}
                extracted_row.update(item)  # Agrega todas las claves y valores de 'item' al diccionario
                extracted_rows.append(extracted_row)

    # Crea un DataFrame a partir de la lista de datos desanidados
    df_new = pd.DataFrame(extracted_rows)
    return df_new

# Nombre de la columna a desanidar
columna = "reviews"

# Llama a la función con el DataFrame df_reviews
df_reviews_desanidado = extraer_values(df_reviews, columna)

df_reviews_desanidado.head() # Mostramos los primeros 5 registros del dataframe.

La columna 'posted' contiene la fecha en la que se publicó la review. Todas las fechas son diferentes, entonces creamos una función que se encargue de convertir la fecha a un formato más legible.

# Cambiamos el formato de fecha de la columna posted.
date_string = df_reviews_desanidado['posted'] # Obtiene la columna "posted"

def transform_date(date_string): # Función para transformar la fecha
    try:
        if isinstance(date_string, str):
            # Elimina "Posted" del principio y analiza la fecha
            date_string = date_string.replace('Posted', '').strip()
            parsed_date = parser.parse(date_string, fuzzy=True)
            return parsed_date
    except ValueError:
        pass  # Ignora fechas incorrectas
    return None  # Devuelve None para fechas inválidas o nulas

# Aplica la función de transformación a la columna "posted"
df_reviews_desanidado['posted'] = df_reviews_desanidado['posted'].apply(transform_date)

# Elimina las filas con valores nulos en la columna 'posted'
df_reviews_final = df_reviews_desanidado.dropna(subset=['posted'])

# Muestra el DataFrame resultante
df_reviews_final.head() 

# Elimina las columnas que no se utilizarán
df_reviews_final.drop(['funny', 'last_edited', 'helpful'], axis=1, inplace=True)

# Cuenta la cantidad de valores nulos en la columna 'posted'
cantidad_valores_nulos = df_reviews_final['posted'].isnull().sum()
print("Cantidad de valores nulos en 'posted':", cantidad_valores_nulos) 

df_reviews_final.info()

# Guardar el dataframe en un archivo csv en la carpeta Clean
df_reviews_final.to_csv('../Datasets/Clean/user_reviews.csv.gz', compression='gzip', index=False, encoding='utf-8')

# Guardar el dataframe en un archivo parquet en la carpeta Datasets
df_reviews_final.to_parquet('../Datasets/user_reviews.parquet.gz', compression='gzip', index=False)
