==> ETL/4.ETL_Final.ipynb <==
import sys
sys.path.append("../Lib")
import utils as ut
import pandas as pd
from textblob import TextBlob
import warnings
warnings.filterwarnings("ignore")

# Cargando datos limpios
df_reviews = pd.read_parquet('../Datasets/user_reviews.parquet.gz')
df_games = pd.read_parquet('../Datasets/steam_games.parquet.gz')
df_items = pd.read_parquet('../Datasets/users_items.parquet.gz')


Vamos a crear una nueva columna llamada 'analisis_sentimiento', reemplazando la columna 'reviews_review'. Realizaremos un análisis de sentimientos de los comentarios en la columna 'reviews', clasificándolos de la siguiente manera: 0 si es negativo, 1 si es neutral o no hay reseña, y 2 si es positivo.

# Aplicar la función de análisis de sentimientos
df_reviews['sentiment_analysis'] = df_reviews['review'].apply(lambda x: TextBlob(str(x)).sentiment.polarity if pd.notna(x) and x != 'No data' else None)

def map_sentiment(score):
    if score is None:
        return 1  # Si la reseña está ausente o 'Sin datos', toma el valor de 1
    elif score < 0:
        return 0  # Malo
    elif score == 0:
        return 1  # Neutral
    else:
        return 2  # Positivo

# Aplicar la función de mapeo a la columna 'sentiment_analysis'
df_reviews['sentiment_analysis'] = df_reviews['sentiment_analysis'].apply(map_sentiment)

### PlayTimeGenre 
Crearemos un dataframe auxiliar que nos muestra género, año, horas reproducidas por género.

playtime_item = df_items[['playtime_forever', 'item_id']]
playtime_item

# Extraiga las columnas 'item_id', 'genres' y 'release_year' en el marco de datos df_games
df_genre_year=df_games[['item_id', 'genres', 'release_year']]
df_genre_year.head()

Uniendo los marcos de datos df_genre_year y playtime_item y obtenga el marco de datos df_playtime_genre para nuestra función futura

df_playtime_genre=df_genre_year.merge(playtime_item, on='item_id')
df_playtime_genre

# Agrupa filas por géneros y años y suma el tiempo jugado
playtime_genre = df_playtime_genre.groupby(['genres', 'release_year'])['playtime_forever'].sum().reset_index()
playtime_genre

### UserForGenre
Usuario por género Crearemos un dataframe auxiliar que nos muestra género, año, usuario, horas jugadas por usuario por año en cada género.

user_item = df_items[['playtime_forever', 'user_id', 'item_id']]
user_item.head()

user_for_genre=df_genre_year.merge(user_item, on='item_id')
user_for_genre

Suponemos que las horas están en minutos (ya que hay valores por año que superan con creces el número de horas de un año) y convertimos el tiempo a horas.

user_for_genre['playtime_hours'] = user_for_genre['playtime_forever'] / 60
user_for_genre['playtime_hours'] = user_for_genre['playtime_hours'].round(2)

user_for_genre = user_for_genre.groupby(['genres', 'user_id', 'release_year'])['playtime_hours'].sum().reset_index()
user_for_genre

Para ahorrar memoria en funciones y modelos futuros, tomamos una muestra aleatoria de datos.

random_rows = 3000000
user_for_genre = user_for_genre.sample(n=random_rows, random_state=42)
user_for_genre

user_for_genre = user_for_genre.groupby(['genres', 'user_id', 'release_year'])['playtime_hours'].sum().reset_index()
user_for_genre

### UsersRecommend
Crearemos un marco de datos auxiliar que nos muestre las columnas item_name recomendadas publicadas y sentiment_analysis para obtener las funciones de recomendación.

it_recommend = df_games[['item_id', 'item_name']]
it_recommend

u_recommend = df_reviews[['item_id', 'posted', 'recommend', 'sentiment_analysis']]
u_recommend

user_recommend = it_recommend.merge(u_recommend, on='item_id')
user_recommend

### sentiment_analysis
Ahora vamos a crear un marco de datos auxiliar donde podremos ver el análisis de sentimiento por año.

release_year = df_games[['item_id', 'release_year']]
release_year = release_year[release_year['release_year'] != 'no data']
release_year

sentiment = df_reviews[['item_id', 'sentiment_analysis']]
sentiment

sentiment_year = release_year.merge(sentiment, on='item_id')
sentiment_year = sentiment_year[['release_year', 'sentiment_analysis']]
sentiment_year['release_year'] = pd.to_numeric(sentiment_year['release_year'], errors='coerce')
sentiment_year

Grabando resultado de las funciones requeridas a formato "csv" y "parquet"

dfs = [playtime_genre, user_for_genre, user_recommend, sentiment_year, df_reviews]
# Nombres correspondientes a cada DataFrame
names = ['playtime_genre', 'user_for_genre', 'user_recommend','sentiment_year', 'reviews_clean']

ut.save_to_csv(dfs, names)

ut.save_to_pq(dfs, names)
